requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
passlib
bcrypt==4.0.1
python-jose[cryptography]
alembic
python-multipart
redis
pydantic[email]